Disallow duplicate link titles
This change came about as a result of the realization that we are dropping numerous client API calls as a part of client generation, both here in schematic and in the ruby version of this tool, Heroics.

While a number of PRs have been initiated to fix the underlying heroku platform api schema itself, these two tools should raise an error and should not proceed to generate an incomplete client in the event of duplicate link titles.

During the discussion here (https://github.com/heroku/api/pull/7787#issuecomment-299169974) we are attempting to do the following:
- fix all of the (sub)schemas of the platform API.
- update schematic and heroics to throw an error when duplicates exist.
- regenerate the Golang and ruby clients, incrementing the major version in order to reflect the breaking changes (title changes -> method name changes) that will be introduced.